# map: reduce allocations for batch operations